Refactor into app_tabs package
