+++ app_tabs/__init__.py
@@ -1,0 +1,3 @@
+from .main_app import main
+
+__all__ = ["main"]
